+++ src/Bridges/NetteDI/TranslationExtension.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Nexendrie\Translation\Bridges\NetteDI;
 
+use Latte\Engine;
+use Latte\Essential\TranslatorExtension;
 use Nette\DI\CompilerExtension;
 use Nette\DI\Definitions\FactoryDefinition;
 use Nette\PhpGenerator\ClassType;
@@ -62,6 +64,8 @@ final class TranslationExtension extends CompilerExtension
     public const SERVICE_ORIGINAL_LOADER = "originalLoader";
     /** @internal */
     public const SERVICE_MESSAGE_SELECTOR = "messageSelector";
+    /** @internal */
+    public const SERVICE_LATTE_EXTENSION = "latteExtension";
 
     private array $resolvers = [
         "environment" => EnvironmentLocaleResolver::class,
@@ -269,14 +273,20 @@ final class TranslationExtension extends CompilerExtension
         if ($builder->hasDefinition($latteFactoryService)) {
             /** @var FactoryDefinition $latteFactory */
             $latteFactory = $builder->getDefinition($latteFactoryService);
-            $latteFactory->getResultDefinition()->addSetup(
-                "addFilter",
-                ["translate", ["@" . $this->prefix(self::SERVICE_TRANSLATOR), "translate"]]
-            );
-            $latteFactory->getResultDefinition()->addSetup(
-                "addProvider",
-                ["translator", "@" . $this->prefix(self::SERVICE_TRANSLATOR)]
-            );
+            if (version_compare(Engine::VERSION, "3", "<")) {
+                $latteFactory->getResultDefinition()->addSetup(
+                    "addFilter",
+                    ["translate", ["@" . $this->prefix(self::SERVICE_TRANSLATOR), "translate"]]
+                );
+                $latteFactory->getResultDefinition()->addSetup(
+                    "addProvider",
+                    ["translator", "@" . $this->prefix(self::SERVICE_TRANSLATOR)]
+                );
+            } else {
+                $latteExtension = $builder->addDefinition($this->prefix(self::SERVICE_LATTE_EXTENSION))
+                    ->setFactory(TranslatorExtension::class, ["@" . $this->prefix(self::SERVICE_TRANSLATOR)]);
+                $latteFactory->getResultDefinition()->addSetup("addExtension", [$latteExtension,]);
+            }
         }
     }